lsm/db: extract db_test_base test fixture
version_set_test.cc carried two near-identical fixtures (VersionSetTest
and CompactionTest) whose persistence/table_cache/version_set wiring,
TearDown, add_sst, and add_file were copy-pasted. An upcoming
compaction_task test needs the same scaffolding plus a fault-injecting
persistence, so pull the shared parts into a header-only db_test_base.

The base owns the options, persistence, table_cache, and version_set,
and exposes add_sst (builds a real SST) and add_file (metadata only).
The data persistence is injected via the constructor (defaulting to the
in-memory layer) so a test can substitute its own. No behavior change;
version_set_test's two fixtures now just configure options and add their
test-specific helpers.
