+++ src/v/lsm/db/tests/BUILD
@@ -123,16 +123,13 @@ redpanda_cc_gtest(
         "version_set_test.cc",
     ],
     deps = [
+        ":db_test_base",
         "//src/v/base",
         "//src/v/lsm/core/internal:files",
         "//src/v/lsm/core/internal:keys",
         "//src/v/lsm/core/internal:options",
-        "//src/v/lsm/db:table_cache",
         "//src/v/lsm/db:version_edit",
         "//src/v/lsm/db:version_set",
-        "//src/v/lsm/io:memory_persistence",
-        "//src/v/lsm/sst:block_cache",
-        "//src/v/lsm/sst:builder",
         "//src/v/test_utils:gtest",
         "@googletest//:gtest",
         "@seastar",
@@ -211,6 +208,28 @@ redpanda_test_cc_library(
     hdrs = ["immutable_tree.h"],
     deps = [
         "//src/v/base",
+        "@seastar",
+    ],
+)
+
+redpanda_test_cc_library(
+    name = "db_test_base",
+    hdrs = ["db_test_base.h"],
+    deps = [
+        "//src/v/base",
+        "//src/v/bytes:iobuf",
+        "//src/v/lsm/core:probe",
+        "//src/v/lsm/core/internal:files",
+        "//src/v/lsm/core/internal:keys",
+        "//src/v/lsm/core/internal:options",
+        "//src/v/lsm/db:table_cache",
+        "//src/v/lsm/db:version_edit",
+        "//src/v/lsm/db:version_set",
+        "//src/v/lsm/io:memory_persistence",
+        "//src/v/lsm/io:persistence",
+        "//src/v/lsm/sst:block_cache",
+        "//src/v/lsm/sst:builder",
+        "@googletest//:gtest",
         "@seastar",
     ],
 )
